+++ src/fct/mult.c
@@ -1,5 +1,6 @@
 #include <strings.h>
 #include <string.h>
+#include <math.h>
 
 #include "mem.h"
 #include "reg.h"
@@ -21,7 +22,7 @@ int fct_mult(inst k, mem memory, registre* reg)
 {
 	long long q=reg[k.r.rs].val*reg[k.r.rt].val;
 	reg[HI].val=(q>>32)&0xffffffff;
-	reg[LO].val=q&0xffffffff;
+	reg[LO].val=((long) q)&0xffffffff;
 	//reg[HI].val=q&0x0000ffff;
 	//reg[LO].val=q&0xffff0000;
 	reg[PC].val+=4;
